+++ src/index.js
@@ -1,4 +1,3 @@
-import { createRequire } from "node:module";
 import core from "./main/core.js";
 import { getSupportInfo as getSupportInfoWithoutPlugins } from "./main/support.js";
 import getFileInfoWithoutPlugins from "./common/get-file-info.js";
@@ -25,8 +24,6 @@ import getLast from "./utils/get-last.js";
 import partition from "./utils/partition.js";
 import { isNonEmptyArray } from "./common/util.js";
 
-const require = createRequire(import.meta.url);
-const { version } = require("../package.json");
 const builtinPlugins = Object.values(languages);
 
 /**
@@ -103,7 +100,6 @@ const debugApis = {
 };
 
 export {
-  version,
   formatWithCursor,
   format,
   check,
@@ -117,3 +113,4 @@ export {
 };
 export * as util from "./common/util-shared.js";
 export * as doc from "./document/index.js";
+export { default as version } from "./main/version.evaluate.js";
