+++ src/main/version.evaluate.js
@@ -1,0 +1,6 @@
+import { createRequire } from "node:module";
+
+const require = createRequire(import.meta.url);
+const { version } = require("../../package.json");
+
+export default version;
